3 things...
changed the return type of template->fields when no fields are present. Returning an array ref or undef is insonsistent and requires all calling code to use an extra check, so I now just return an array ref (with zero length if there are no fields). If people only care about weather there are fields, one check is needed as before. If people only care about processing the field list, loop code should work when there are no fields in the template.

changed template->field to a getter / setter (was just a getter). This works nicely :-D ... although I'm not sure how I was expecting it to fail... TODO: add methods to add and remove fields

hacked in functionality to 'delete' a template ... I'm sure there is a better way to do it :-(